float 32 and prints

+++ src/pyVertexModel/Kg/kg.py
@@ -8,8 +8,8 @@ class Kg:
 
     def __init__(self, Geo):
         self.dimg = (Geo.numY + Geo.numF + Geo.nCells) * 3
-        self.g = np.zeros(self.dimg, dtype=np.float16)
-        self.K = np.zeros([self.dimg, self.dimg], dtype=np.float16)
+        self.g = np.zeros(self.dimg, dtype=np.float32)
+        self.K = np.zeros([self.dimg, self.dimg], dtype=np.float32)
         self.energy = None
         self.dim = 3
